% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby.R
\name{sortby}
\alias{sortby}
\title{creates a function to sort
df is the data frame
col is the columns}
\usage{
sortby(df, col, desc = FALSE)
}
\arguments{
\item{df}{is dataframe}

\item{col}{is columns}

\item{desc}{order in which data is sorted}
}
\description{
creates a function to sort
df is the data frame
col is the columns
}
